Added additional configuration Started to clean code up by deleting some interfaces and classes Put all classes into their own file Still a lot of clean up to do.
