+++ ABus.AzureServiceBus/TransportInstance.cs
@@ -1,0 +1,13 @@
+using ABus.Contracts;
+using Microsoft.ServiceBus;
+
+namespace ABus.AzureServiceBus
+{
+    internal class TransportInstance : TransportDefinition
+    {
+        public NamespaceManager Namespace { get; set; }
+
+        public string ConnectionString { get { return string.Format("Endpoint={0};{1}", this.Uri, this.Credentials); } }
+
+    }
+}
