+++ ABus.Contracts/ABus.Contracts.csproj
@@ -40,18 +40,18 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DefaultTransactionManager.cs" />
+    <Compile Include="HostConfiguration.cs" />
     <Compile Include="IAssemblyResolver.cs">
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="IBus.cs" />
-    <Compile Include="ICommand.cs" />
-    <Compile Include="IConfigureHost.cs" />
-    <Compile Include="IEvent.cs" />
     <Compile Include="IHandleMessage.cs" />
-    <Compile Include="IMessage.cs" />
+    <Compile Include="IManageOutboundMessages.cs" />
+    <Compile Include="IManageTransactions.cs" />
     <Compile Include="IMessageQueue.cs" />
     <Compile Include="IMessageTransport.cs" />
-    <Compile Include="IUnitOfWork.cs" />
+    <Compile Include="OutboundMessageManager.cs" />
     <Compile Include="MetaData.cs" />
     <Compile Include="MetaDataCollection.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
