--- ABus.Contracts/ICommand.cs
@@ -1,6 +1,0 @@
-namespace ABus.Contracts
-{
-    public interface ICommand : IMessage
-    {
-    }
-}
