+++ ABus.Sample/ABus.Sample.csproj
@@ -39,8 +39,10 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Class1.cs" />
+    <Compile Include="SampleMessageHandler.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TestMessage2Event.cs" />
+    <Compile Include="TestMessageCommand.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\ABus.Contracts\ABus.Contracts.csproj">
