+++ ABus.Sample/TestMessage2Event.cs
@@ -1,0 +1,9 @@
+namespace ABus.Sample
+{
+    public class TestMessage2Event
+    {
+        public string Name { get; set; }
+
+        public string Addresss { get; set; }
+    }
+}
