+++ ABus.Sample/TestMessageCommand.cs
@@ -1,0 +1,9 @@
+namespace ABus.Sample
+{
+    public class TestMessageCommand
+    {
+        public string Name { get; set; }
+
+        public string Addresss { get; set; }
+    }
+}
