+++ ABus.Unity/UnityContainerAdaptor.cs
@@ -1,30 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
-using ABus.Contracts;
 using Microsoft.Practices.ServiceLocation;
 using Microsoft.Practices.Unity;
 
 namespace ABus.Unity
 {
-    public class UnityBootstraper : UnityContainerAdaptor, IABusContainer
-    {
-        public UnityBootstraper(IUnityContainer container) : base(container)
-        {
-            this.RegisterTypes();
-        }
-
-        public UnityBootstraper()
-        {
-            this.RegisterTypes();
-        }
- 
-        void RegisterTypes()
-        {
-            this.container.RegisterType<IAssemblyResolver, AssemblyResolver>(new ContainerControlledLifetimeManager());
-            this.container.RegisterType<IManageTransactions, DefaultTransactionManager>();
-        }
-    }
-
     public class UnityContainerAdaptor : ServiceLocatorImplBase
     {
         protected IUnityContainer container;
