+++ ABus/Config/Transport/TransportDefinitionGrammar.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Reflection.Emit;
 using ABus.Contracts;
-using Microsoft.Practices.Unity;
 
 namespace ABus.Config.Transport
 {
