+++ ABus/RegisteredHandler.cs
@@ -1,17 +1,8 @@
 using System;
-using System.Collections.ObjectModel;
 using System.Reflection;
 
 namespace ABus
 {
-    public class RegisteredHandlerCollection : KeyedCollection<string, RegisteredHandler>
-    {
-        protected override string GetKeyForItem(RegisteredHandler item)
-        {
-            return item.SubscriptionName;
-        }
-    }
-
     public class RegisteredHandler
     {
         /// <summary>
