+++ ABus/RegisteredMessageType.cs
@@ -1,17 +1,8 @@
 using System;
-using System.Collections.ObjectModel;
 using ABus.Contracts;
 
 namespace ABus
 {
-    public class RegisteredMessageTypeCollection : KeyedCollection<string, RegisteredMessageType>
-    {
-        protected override string GetKeyForItem(RegisteredMessageType item)
-        {
-            return item.FullName;
-        }
-    }
-
     public class RegisteredMessageType
     {
         /// <summary>
