+++ ABus/StartupStages.cs
@@ -1,0 +1,7 @@
+namespace ABus
+{
+    public class StartupStages
+    {
+        public const string Initialize = "Initialize";
+    }
+}
