+++ ABus/Tasks/Startup/ValidateQueuesTask.cs
@@ -12,7 +12,7 @@ namespace ABus.Tasks.Startup
                 // get message transport
                 var transport = context.TransportInstances[m.Transport.Name];
                 var endpoint = new QueueEndpoint {Host = m.Transport.Uri, Name = m.Queue};
-                var exists = transport.QueueExists(endpoint).Result;
+                var exists = transport.QueueExistsAsync(endpoint).Result;
                 if (exists)
                     context.Trace.Verbose(string.Format("Found: queue: {0}:{1}", m.Transport.Name, m.Queue));
                 else
@@ -21,7 +21,7 @@ namespace ABus.Tasks.Startup
                     if (context.Configuration.EnsureQueuesExist)
                     {
                         // Create queue
-                        transport.CreateQueue(endpoint).Wait();
+                        transport.CreateQueueAsync(endpoint).Wait();
                         context.Trace.Information(string.Format("Created: queue: {0}:{1}", m.Transport.Name, m.Queue));
                     }
                 }
